pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "fraud-detection"
version = "1.0.0"
description = "Production fraud detection system — BAF Suite Variant II"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.35.0",
    "sagemaker>=2.232.0",
    "mlflow>=2.17.0",
    "xgboost>=2.1.0",
    "scikit-learn>=1.5.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "polars>=1.15.0",
    "shap>=0.46.0",
    "great-expectations>=1.2.0",
    "evidently>=0.4.0",
    "fastapi>=0.115.0",
    "mangum>=0.19.0",
    "uvicorn>=0.32.0",
    "pydantic>=2.9.0",
    "pyyaml>=6.0",
    "joblib>=1.4.0",
    "matplotlib>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "moto[s3,sagemaker]>=5.0.0",
    "httpx>=0.27.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"

[tool.setuptools.packages.find]
include = ["src*", "app*"]
